+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 
 - FIXES: regarding array initialisation
 - ADDITIONS: decaying arrays to pointers in function calls
+- FIXES: handle void[] and returning stack-allocated arrays
 
 ## TODO